// packages/components/src/ui/Card/CardSection.styles.d.ts
export declare const useStyles: () => {
  cardSection: {
    display: 'flex';
    flexDirection: 'column';
    gap: import('@alveole/theme').Spacing;
    width: '100%';
  };
  title: {
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    gap: import('@alveole/theme').Spacing;
  };
  titleIcon: {
    display: 'flex';
    alignItems: 'center';
  };
  titleText: {
    color: '#151617';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
  };
  description: {
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    gap: import('@alveole/theme').Spacing;
  };
  descriptionIcon: {
    display: 'flex';
    alignItems: 'center';
  };
  descriptionText: {
    color: '#373A3F';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
  };
  disabledText: {
    color: '#8D97AC';
  };
};
//# sourceMappingURL=CardSection.styles.d.ts.map